requirements.txt
# Georgia Algorithmic Framing Study
# Python 3.9+ Dependencies

# Data manipulation
pandas>=1.3.5
numpy>=1.21.5

# Statistical analysis
scipy>=1.7.3
statsmodels>=0.13.2
linearmodels>=4.20

# Natural Language Processing
nltk>=3.6.7
textblob>=0.17.1
vaderSentiment>=3.3.2      # Preferred over TextBlob for social media text
gensim>=4.1.2

# Machine learning utilities
scikit-learn>=0.24.2

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2

# Data collection
requests>=2.26.0
census>=0.8.19             # Census API wrapper

# Configuration
pyyaml>=6.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.6
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
tqdm>=4.62.3
python-dateutil>=2.8.2
